==> Processing/In-house data processing/Annotation_process/05_ann_lvl2/05_redoing_metabolic_annotations.rmd <==
# Tomorrow,
###  1. Check with Juan the markers of the objects to really confirm my findings.
###  2. Overlap all of the previous annotations in the bigger objects.
###  3. Compare the findings of annotation with the results in wetlab, do things correlate with the values found?
---
author: "Ivo Borkus"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:  
  pdf_document: default
  html_document:
    toc: true
    toc_depth: 2
title: "`r paste0('Immune compartment analysis of singlecellRNAseq samples')`"
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "/scratch_isilon/groups/singlecell/iborkus/Projects/brain_met") # nolint
knitr::opts_chunk$set(warning = FALSE, tidy.opts = list(width.cutoff = 60), tidy = FALSE) # nolint
set.seed(91774)
# install.packages("languageserver")
```

```{R general functions, eval =T}
library(here)
setwd("/scratch_isilon/groups/singlecell/iborkus/Projects/brain_met")
here::i_am("02_scripts/01_phase/lineage_annotation/per_sample_annotation.rmd")
output_figs <- "03_processing/13_phase_6/figs/metabolic/"
meta_data <- read.csv(here("03_processing/01_QC/data/meta_data.csv"))
for (sample in meta_data$sample_name) {
    print(sample)
}
source("02_scripts/04_integration/functions.R")
source("02_scripts/06_analysing_annotations/utils.R")
library(dittoSeq)


```
```{R saving object,eval = T}
tseur <- readRDS( file = "03_processing/12_phase_5/data/Tcell_seurat_annotated.rds")
reduction_name <- "Tcell"

Idents(object = tseur) <- "RNA_snn_res.0.9"# Order is messed up, so the order of sizes does not correspond
# new.cluster.ids <- c("Exhausted_CD8", "Tcells_CD8", "Bad-Quality", "Unknown", "Unknown", "Proliferative_1", "Tcells_CD4", "Tcells_Naive", "Tregs", "Tregs_2", "NKs")
new.cluster.ids <- c("Exhausted_CD8", "Proliferating_CD8", "Tcells_CD8", "Unknown", "Unknown", "Proliferative_Tcells", "Tcells_CD4_Naive", "Unknown", "Tregs", "Tcells_CD4", "Tcells_CD4_proliferating","Tregs_2","NK")
new.cluster.ids_2 <- c("Exhausted_CD8", "Proliferating_CD8", "Tcells_CD8", "Unknown_1", "Unknown_2", "Proliferative_Tcells", "Tcells_CD4_Naive", "Unknown_3", "Tregs", "Tcells_CD4", "Tcells_CD4_proliferating","Tregs_2","NK")
names(new.cluster.ids_2) <- levels(tseur)
tseur <- RenameIdents(tseur, new.cluster.ids_2 )
tseur[["ann_lvl_2_Tcells_specific"]] <- Idents(tseur)
Idents(object = tseur) <- "RNA_snn_res.0.9"
names(new.cluster.ids) <- levels(tseur)
tseur <- RenameIdents(tseur, new.cluster.ids)
tseur[["ann_lvl_2_Tcells"]] <- Idents(tseur)
DimPlot(tseur, reduction = paste0("umap.harmony.",reduction_name),pt.size = 0.1, label = T, group.by = "ann_lvl_1_filtered") & NoAxes()
DimPlot(tseur, reduction = paste0("umap.harmony.",reduction_name),pt.size = 0.1, label = T, group.by = "ann_lvl_2_Tcells") & NoAxes()

dittoDimPlot(tseur, reduction.use = paste0("umap.harmony.",reduction_name), var = "ann_lvl_2_Tcells_specific", do.label = TRUE)
```



```{R selecting the high_mt cells}

Idents(object = tseur) <- "RNA_snn_res.0.05"

dittoDimPlot(tseur, reduction.use = paste0("umap.harmony.",reduction_name), var = "RNA_snn_res.0.05", do.label = TRUE)
ggsave(filename=paste0(output_figs,"umap_mt_difference.png"), width = 10, height = 10)  

Idents(tseur) %>% unique()
celltypes <-  c("1")

low_mt_tseur <- subset(x = tseur, idents = celltypes, invert = F)
high_mt_tseur <- subset(x = tseur, idents = celltypes, invert = T)

```

```{R lisi function}
ndims = 25
lisi_running <- function(seurat_obj, reduction_name, ndims, batch_col) {
    Embeddings(seurat_obj, reduction = paste0("harmony_pca.", reduction_name))[,1:ndims] %>%  compute_lisi( ,meta_data = seurat_obj@meta.data,label_colnames =  c(batch_col))-> lis
    Embeddings(seurat_obj, reduction = paste0("pca_", reduction_name))[,1:ndims] %>%  compute_lisi( ,meta_data = seurat_obj@meta.data,label_colnames =  c(batch_col))-> lis_no_int
    lis_no_int$integrated <- "no"
    lis$integrated <- "yes"
    combined <- rbind(lis_no_int, lis)
    t.test(sample ~ integrated, data = combined) -> test_result
    plot <- ggplot(combined, mapping = aes(x = integrated, y = sample)) +
        geom_boxplot() +
        theme_minimal() & labs(title = paste0("Lisi score pre and post integration with a significand t-test statistic: ", round(test_result$statistic,2)))
    ggsave(plot, filename = paste0(output_figs, "Lisi_samples_",reduction_name,".png"))
    return(plot)

}

excel_sheet <- function(markers, output_dir, name) {
    library(writexl)
    print(paste0("Output will be put in: ", output_dir, name, ".xlsx"))
    if (file.exists(output_dir)) {
        markers %>%
            arrange(cluster, desc(avg_log2FC)) %>% # Arrange within each cluster
            group_by(cluster) %>%
            select(cluster, pct.1, pct.2, p_val, p_val_adj, avg_log2FC, gene) %>%
            group_split() %>% # Split into list by 'cluster'
            setNames(unique(markers$cluster)) %>% # Name list elements
            writexl::write_xlsx(paste0(output_dir, name, ".xlsx"))
    } else {
        stop("Directory does not exist")
    }
}
output_excel <- "03_processing/13_phase_6/data/metabolic/"

```

```{R analysis for high_mt}


ncol(high_mt_tseur)
reduction_name <- "high_mt"

high_mt_tseur[["RNA"]] <- split(high_mt_tseur[["RNA"]], f = high_mt_tseur$sample)
high_mt_tseur <- FindVariableFeatures(high_mt_tseur, selection.method = "vst", nfeatures = 2000, verbose = TRUE)
hvf_info <- HVFInfo(high_mt_tseur)
top_variable_genes <- hvf_info[order(hvf_info$variance.standardized, decreasing = TRUE), ]
top_2000_genes <- rownames(top_variable_genes)[1:2000]
VariableFeatures(high_mt_tseur) <- top_2000_genes
high_mt_tseur <- ScaleData(object = high_mt_tseur, features = rownames(high_mt_tseur), verbose = FALSE)
high_mt_tseur <- RunPCA(
    object = high_mt_tseur, nfeatures.print = 5, ndims.print = 1:2, reduction.name = paste0("pca_", reduction_name)
)
ElbowPlot(high_mt_tseur, reduction =  paste0("pca_", reduction_name),ndims = 50)

VizDimLoadings(high_mt_tseur,
    dims = 1:5, reduction =  paste0("pca_", reduction_name), nfeatures = 15
)
# PCA_elbow <- ElbowPlot(high_mt_tseur, reduction =  paste0("pca_", reduction_name),ndims = 50)
# Genes_influence_PCA <- VizDimLoadings(high_mt_tseur,
#     dims = 1:5, reduction =  paste0("pca_", reduction_name), nfeatures = 15
# )
# ggsave(PCA_elbow, filename = paste0(output_figs, reduction_name, "_PCA_elbow_1.png"), width = 10, height= 10)
# ggsave(Genes_influence_PCA, filename = paste0(output_figs, reduction_name, "_PCA_loadings_1.png"), width = 15, height = 15)

high_mt_tseur <- IntegrateLayers(
    object = high_mt_tseur, method = HarmonyIntegration,
    orig.reduction = paste0("pca_", reduction_name), new.reduction = paste0("harmony_pca.", reduction_name),
    verbose = TRUE
)

res_values <- c(0.01, 0.05, 0.1, 0.2, 0.3, 0.5, 0.7, 0.9, 1.5)
high_mt_tseur <- FindNeighbors(high_mt_tseur, reduction = paste0("harmony_pca.", reduction_name), dims = 1:25)
high_mt_tseur <- FindClusters(high_mt_tseur, resolution = res_values, algorithm = 4)
high_mt_tseur <- RunUMAP(high_mt_tseur, reduction = paste0("harmony_pca.", reduction_name), dims = 1:25, reduction.name = paste0("umap.harmony.", reduction_name))
Running_plots_post_harmony(high_mt_tseur, reduction_name)
high_mt_tseur <- JoinLayers(high_mt_tseur)
# saveRDS(processed_list, file = "03_processing/09_subset_annotation/data/processed_list.rds")
# high_mt_tseur@meta.data$ann_lvl_1 %>% unique()

lisi_running(seurat_obj = high_mt_tseur,reduction_name = "high_mt",ndims = 25,batch_col = "sample")
dittoDimPlot(high_mt_tseur, reduction.use = paste0("umap.harmony.","Tcell"), var = "RNA_snn_res.0.05", do.label = TRUE)
dittoDimPlot(high_mt_tseur, reduction.use = paste0("umap.harmony.",reduction_name), var = "RNA_snn_res.0.05",split.by = "RNA_snn_res.0.05", do.label = TRUE)
# saveRDS(high_mt_tseur, file = "03_processing/13_phase_6/data/metabolic/high_mt_tseur.rds")

# ggsave(filename=paste0(output_figs,"umap_mt_high.png"), width = 10, height = 10)  
```

```{R picking res}
dittoDimPlot(high_mt_tseur, reduction.use = paste0("umap.harmony.","Tcell"), var = "RNA_snn_res.0.05", do.label = TRUE)
ggsave(filename=paste0(output_figs,"umap_selection_High_mt.png"), width = 10, height = 10)  

FeaturePlot(high_mt_tseur,reduction = paste0("umap.harmony.",reduction_name), features = c("percent_mt")) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_mt_high_specific.png"), width = 10, height = 10)  


Idents(object = high_mt_tseur) <- "RNA_snn_res.0.3"

DimPlot(high_mt_tseur, reduction = paste0("umap.harmony.",reduction_name), label = T) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_0.3_high.png"), width = 10, height = 10)  

VlnPlot(high_mt_tseur,features = c("percent_mt","percent.rb","nCount_RNA","nFeature_RNA","doublet.score"))
ggsave(filename=paste0(output_figs,"violin_basic_stats_0.3_high.png"), width = 10, height = 10) 


```

```{R annotation_high_mt}
table(Idents(high_mt_tseur))

high_mt_tseur.markers <- FindAllMarkers(
    object = high_mt_tseur,
    only.pos = TRUE, min.pct = 0.25,
    logfc.threshold = 0.25
)
excel_sheet(high_mt_tseur.markers, output_excel, "high_mt_tseur_correct.0.3")

FeaturePlot(high_mt_tseur,reduction = paste0("umap.harmony.",reduction_name), features = c("CD4","CD8A","CD8B","CD3D","CD3E","NCAM1","KLRB1","SELL","CCR7","LEF1","TCF7","CXCL13"), order = TRUE) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_features_high_mt.png"), width = 15, height = 10)  
```

```{R analysis for low_mt}

ncol(low_mt_tseur)
reduction_name <- "low_mt"
low_mt_tseur[["RNA"]] <- split(low_mt_tseur[["RNA"]], f = low_mt_tseur$sample)
low_mt_tseur <- FindVariableFeatures(low_mt_tseur, selection.method = "vst", nfeatures = 2000, verbose = TRUE)
hvf_info <- HVFInfo(low_mt_tseur)
top_variable_genes <- hvf_info[order(hvf_info$variance.standardized, decreasing = TRUE), ]
top_2000_genes <- rownames(top_variable_genes)[1:2000]
VariableFeatures(low_mt_tseur) <- top_2000_genes
low_mt_tseur <- ScaleData(object = low_mt_tseur, features = rownames(low_mt_tseur), verbose = FALSE)
low_mt_tseur <- RunPCA(
    object = low_mt_tseur, nfeatures.print = 5, ndims.print = 1:2, reduction.name = paste0("pca_", reduction_name)
)
ElbowPlot(low_mt_tseur, reduction =  paste0("pca_", reduction_name),ndims = 50)
VizDimLoadings(low_mt_tseur,
    dims = 1:5, reduction =  paste0("pca_", reduction_name), nfeatures = 15
)
# PCA_elbow <- ElbowPlot(low_mt_tseur, reduction =  paste0("pca_", reduction_name),ndims = 50)
# Genes_influence_PCA <- VizDimLoadings(low_mt_tseur,
#     dims = 1:5, reduction =  paste0("pca_", reduction_name), nfeatures = 15
# )
# ggsave(PCA_elbow, filename = paste0(output_figs, reduction_name, "_PCA_elbow_1.png"), width = 10, height= 10)
# ggsave(Genes_influence_PCA, filename = paste0(output_figs, reduction_name, "_PCA_loadings_1.png"), width = 15, height = 15)

low_mt_tseur <- IntegrateLayers(
    object = low_mt_tseur, method = HarmonyIntegration,
    orig.reduction = paste0("pca_", reduction_name), new.reduction = paste0("harmony_pca.", reduction_name),
    verbose = TRUE
)

res_values <- c(0.01, 0.05, 0.1, 0.2, 0.3, 0.5, 0.7, 0.9, 1.5)
low_mt_tseur <- FindNeighbors(low_mt_tseur, reduction = paste0("harmony_pca.", reduction_name), dims = 1:25)
low_mt_tseur <- FindClusters(low_mt_tseur, resolution = res_values, algorithm = 4)
low_mt_tseur <- RunUMAP(low_mt_tseur, reduction = paste0("harmony_pca.", reduction_name), dims = 1:25, reduction.name = paste0("umap.harmony.", reduction_name))
Running_plots_post_harmony(low_mt_tseur, reduction_name)
low_mt_tseur <- JoinLayers(low_mt_tseur)
# saveRDS(processed_list, file = "03_processing/09_subset_annotation/data/processed_list.rds")
# low_mt_tseur@meta.data$ann_lvl_1 %>% unique()

lisi_running(seurat_obj = low_mt_tseur,reduction_name = "low_mt",ndims = 25,batch_col = "sample")
dittoDimPlot(low_mt_tseur, reduction.use = paste0("umap.harmony.","Tcell"), var = "ann_lvl_2_Tcells_specific", do.label = TRUE)
ggsave(filename=paste0(output_figs,"umap_mt_low.png"), width = 10, height = 10)  


# saveRDS(low_mt_tseur, file = "03_processing/13_phase_6/data/metabolic/low_mt_tseur.rds")

```


```{R picking res low-mt}

dittoDimPlot(low_mt_tseur, reduction.use = paste0("umap.harmony.","Tcell"), var = "RNA_snn_res.0.05", do.label = TRUE)
ggsave(filename=paste0(output_figs,"umap_selection_low_mt.png"), width = 10, height = 10)  


FeaturePlot(low_mt_tseur,reduction = paste0("umap.harmony.",reduction_name), features = c("percent_mt")) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_mt_low_specific.png"), width = 10, height = 10)  


Idents(object = low_mt_tseur) <- "RNA_snn_res.0.5"

DimPlot(low_mt_tseur, reduction = paste0("umap.harmony.",reduction_name), label = T) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_low_0.5.png"), width = 10, height = 10)  

VlnPlot(low_mt_tseur,features = c("percent_mt","percent.rb","nCount_RNA","nFeature_RNA","doublet.score"))
ggsave(filename=paste0(output_figs,"violin_basic_stats_low_0.5.png"), width = 10, height = 10) 



```



```{R annotation_low_mt}

table(Idents(low_mt_tseur))
low_mt_tseur.markers <- FindAllMarkers(
    object = low_mt_tseur,
    only.pos = TRUE, min.pct = 0.25,
    logfc.threshold = 0.25
)
excel_sheet(low_mt_tseur.markers, output_excel, "low_mt_tseur.0.5_correct")

FeaturePlot(low_mt_tseur,reduction = paste0("umap.harmony.",reduction_name), features = c("CD4","CD8A","CD8B","CD3D","CD3E","NCAM1","KLRB1","SELL","CCR7","LEF1","TCF7","CXCL13"), order = TRUE) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_features_low_mt.png"), width = 15, height = 10)  


FeaturePlot(low_mt_tseur,reduction = paste0("umap.harmony.",reduction_name), features = c(
    "CD4","CD8A","CD8B","CX3CR1","DUSP2","XCL1","LYAR","NCR3","ZNF683","KLRC2","CD7","TRDV1","FXYD2","CXCR3","GZMB","HLA-DRB1"
    ), order = TRUE) & NoAxes()
ggsave(filename=paste0(output_figs,"umap_features_Specifc_low_mt.png"), width = 15, height = 10)  

colnames(low_mt_tseur@meta.data)
dittoDimPlot(low_mt_tseur, reduction.use = paste0("umap.harmony.","Tcell"), var = "ann_lvl_2_Tcells_specific",split.by = "RNA_snn_res.0.5" ,do.label = TRUE)
dittoDimPlot(low_mt_tseur, reduction.use = paste0("umap.harmony.",reduction_name), var = "ann_lvl_2_Tcells_specific",split.by = "ann_lvl_2_Tcells_specific" ,do.label = TRUE)
ggsave(filename=paste0(output_figs,"Low-mt_previous_annotation_on_new_object.png"), width = 15, height = 10)  


dittoDimPlot(high_mt_tseur, reduction.use = paste0("umap.harmony.","high_mt"), var = "ann_lvl_2_Tcells_specific",split.by = "ann_lvl_2_Tcells_specific" ,do.label = TRUE)
ggsave(filename=paste0(output_figs,"High-mt_previous_annotation_on_new_object.png"), width = 15, height = 10)  


```


```{R naming the objects}
### Low mt
Idents(object = low_mt_tseur) <- "RNA_snn_res.0.5"# Order is messed up, so the order of sizes does not correspond
# Idents(tseur) <- factor(x = Idents(tseur), levels = as.character(sort(as.numeric(levels(tseur)))))
Idents(low_mt_tseur) %>% levels()
new.cluster.ids <- c("Exhausted_CD8", "Proliferating_CD8", "Central_memory_CD4", "Tfh_cells", "Tregs", "NKT_MAIT", "Proliferating_CD4")
names(new.cluster.ids) <- levels(low_mt_tseur)
low_mt_tseur <- RenameIdents(low_mt_tseur, new.cluster.ids )
low_mt_tseur[["low_mt_annotation"]] <- Idents(low_mt_tseur)
dittoDimPlot(low_mt_tseur, reduction.use = paste0("umap.harmony.","low_mt"), var = "low_mt_annotation",split.by = "primary_sample" ,do.label = TRUE)
ggsave(filename=paste0(output_figs,"low_mt_annotated-per-sample.png"), width = 15, height = 10)  


dittoDimPlot(low_mt_tseur, reduction.use = paste0("umap.harmony.","low_mt"), var = "low_mt_annotation",do.label = TRUE)
ggsave(filename=paste0(output_figs,"low_mt_annotated.png"), width = 15, height = 10)  

table(Idents(low_mt_tseur))

```

```{R high_mt}
### high mt
Idents(object = high_mt_tseur) <- "RNA_snn_res.0.3"# Order is messed up, so the order of sizes does not correspond
# Idents(tseur) <- factor(x = Idents(tseur), levels = as.character(sort(as.numeric(levels(tseur)))))
Idents(high_mt_tseur) %>% levels()
new.cluster.ids <- c("Exhausted_CD8", "Proliferating_CD8", "Central_memory_CD4", "Tfh_cells", "Tregs", "NKT_MAIT", "Proliferating_CD4")
names(new.cluster.ids) <- levels(high_mt_tseur)
high_mt_tseur <- RenameIdents(high_mt_tseur, new.cluster.ids )
high_mt_tseur[["high_mt_annotation"]] <- Idents(high_mt_tseur)
dittoDimPlot(high_mt_tseur, reduction.use = paste0("umap.harmony.","high_mt"), var = "high_mt_annotation",split.by = "primary_sample" ,do.label = TRUE)
ggsave(filename=paste0(output_figs,"high_mt_annotated-per-sample.png"), width = 15, height = 10)  


dittoDimPlot(high_mt_tseur, reduction.use = paste0("umap.harmony.","high_mt"), var = "high_mt_annotation",do.label = TRUE)
ggsave(filename=paste0(output_figs,"high_mt_annotated.png"), width = 15, height = 10)  

table(Idents(high_mt_tseur))
```

```{R subcluster}

```
